Lift DPA4 packed SO(3) helpers into shared so3 primitives.
Share indexing, focus linear/gated activation, degree RMSNorm, and Bessel×C³ RBF so DPA4 keeps thin re-exports while EMFA SO(2) and quaternion Wigner-D stay local.
