+++ docs/user_guide/models/architecture_catalog.rst
@@ -77,7 +77,7 @@ External models are declared under :code:`Modelhub.external_FFs` with the same :
 
 **EquiformerV3** (:code:`architecture: equiformer_v3`) is a native port of Liao et al. (2026, arXiv:2604.09130) under :code:`enerzyme/models/equiformer_v3/`. It extends shared :code:`so3/` with merged layer norm, SwiGLU-S² activations, fused SO(2) linears (:code:`SO2Linear`), polynomial attention envelopes, and asymmetric S² grids. The Core keeps the same :code:`atom_feature` / :code:`atom_sphere_feature` contract as eSCN/V2. Default stacks use :code:`SimpleReadout` + :code:`EnergyReduce` + :code:`Force` (DeNS / stress / direct force heads stay outside the Core). Example: :code:`enerzyme/config/equiformer_v3_layers_example.yaml`. Parity: :code:`test/test_equiformer_v3_parity_*.py`. Enerzymette only needs :code:`architecture: equiformer_v3` plus a resolved :code:`config.yaml`.
 
-**DPA4** (:code:`architecture: dpa4`) implements the EMFA SO(2) descriptor of Li et al. (2026, arXiv:2606.02419) under :code:`enerzyme/models/dpa4/`. Its Core owns the Bessel/C³ radial cache, geometric/environment seed, m-truncated SO(2) interactions, and Lebedev-grid equivariant FFN. It emits :code:`atom_feature` / :code:`atom_sphere_feature`, so the standard :code:`SimpleReadout` + :code:`EnergyReduce` + :code:`Force` stack applies. Example: :code:`enerzyme/config/dpa4_layers_example.yaml`. Tests: :code:`test/test_dpa4_core.py`, :code:`test/test_dpa4_parity_ops.py`. Enerzymette only needs :code:`architecture: dpa4` plus a resolved :code:`config.yaml`.
+**DPA4** (:code:`architecture: dpa4`) implements the EMFA SO(2) descriptor of Li et al. (2026, arXiv:2606.02419) under :code:`enerzyme/models/dpa4/`. Shared SO(3) helpers (C³ envelope, Lebedev projection, packed indexing, focus linear / gated activation, degree RMSNorm, Bessel×C³ RBF) live in :code:`enerzyme/models/so3/`; focus-major EMFA SO(2) and quaternion Wigner-D stay local. The Core owns the radial cache, geometric/environment seed, m-truncated interactions, and Lebedev-grid equivariant FFN. It emits :code:`atom_feature` / :code:`atom_sphere_feature`, so the standard :code:`SimpleReadout` + :code:`EnergyReduce` + :code:`Force` stack applies. Example: :code:`enerzyme/config/dpa4_layers_example.yaml`. Tests: :code:`test/test_dpa4_core.py`, :code:`test/test_dpa4_parity_ops.py`. Enerzymette only needs :code:`architecture: dpa4` plus a resolved :code:`config.yaml`.
 
 **So3krates** (:code:`architecture: so3krates`) is a native port of Frank et al. (NeurIPS 2022) under :code:`enerzyme/models/so3krates/`, following the So3krates-torch EuclideanTransformer (fused FeatureBlock + GeometricBlock + InteractionBlock). Shared :code:`RealSphericalHarmonics` and :code:`L0Contraction` live in :code:`enerzyme/models/so3/`. The Core emits :code:`atom_feature` (invariant stream ``x``, :code:`feature_irreps: "Fx0e"`) and :code:`atom_sphere_feature` (SPHC ``χ`` with shape ``[N, m_tot]``, **not** eSCN/EquiformerV2's ``[N, (lmax+1)^2, C]`` — :code:`SphereSampleReadout` does not apply). Default stacks use :code:`BernsteinRBF` + :code:`SimpleReadout` + :code:`EnergyReduce` + :code:`Force`. Long-range physics (ZBL / electrostatics / dispersion) belong in post-core layers, not inside the Core. Example: :code:`enerzyme/config/so3krates_layers_example.yaml`. Parity: :code:`test/test_so3krates_parity_ops.py`.
