+++ enerzyme/models/so3/linear.py
@@ -6,6 +6,7 @@
 from __future__ import annotations
 
 import math
+from typing import Optional
 
 import torch
 from torch.nn import Parameter
@@ -103,4 +104,70 @@ class SO3Linear(torch.nn.Module):
             f"{self.__class__.__name__}(in_features={self.in_features}, "
             f"out_features={self.out_features}, lmax={self.lmax}, "
             f"bias={(self.bias is not None)})"
+        )
+
+
+class SO3FocusLinear(torch.nn.Module):
+    """Degree-wise linear on packed ``(N, D, F, C)`` features (DPA4 / SeZM).
+
+    Weights are shared across ``m`` within each ``l``, with an explicit focus
+    stream axis ``F``. Distinct from :class:`SO3Linear` (``[N, D, C]``,
+    EquiformerV3) and :class:`SO3_LinearV2` (``SO3_Embedding`` wrapper).
+    """
+
+    def __init__(
+        self,
+        lmax: int,
+        in_channels: int,
+        out_channels: int,
+        n_focus: int = 1,
+        mlp_bias: bool = False,
+        init_std: Optional[float] = None,
+    ) -> None:
+        super().__init__()
+        from .indexing import map_degree_idx
+
+        self.lmax = int(lmax)
+        self.in_channels = int(in_channels)
+        self.out_channels = int(out_channels)
+        self.n_focus = int(n_focus)
+        self.mlp_bias = bool(mlp_bias)
+        weight = torch.empty(
+            self.lmax + 1, self.in_channels, self.n_focus * self.out_channels
+        )
+        if init_std is not None:
+            if init_std == 0.0:
+                torch.nn.init.zeros_(weight)
+            else:
+                torch.nn.init.normal_(weight, std=float(init_std))
+        else:
+            torch.nn.init.xavier_uniform_(weight.view(self.lmax + 1, -1))
+        self.weight = Parameter(weight)
+        if self.mlp_bias:
+            self.bias = Parameter(torch.zeros(self.n_focus * self.out_channels))
+        else:
+            self.register_parameter("bias", None)
+        self.register_buffer(
+            "expand_index",
+            torch.as_tensor(map_degree_idx(self.lmax), dtype=torch.long),
+        )
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        # x: (N, D, F, Cin)
+        w = self.weight.view(
+            self.lmax + 1, self.in_channels, self.n_focus, self.out_channels
+        )
+        w = w.index_select(0, self.expand_index)  # (D, Cin, F, Cout)
+        w = w.permute(0, 2, 1, 3)  # (D, F, Cin, Cout)
+        out = torch.matmul(x.unsqueeze(-2), w.unsqueeze(0)).squeeze(-2)
+        if self.bias is not None:
+            bias = self.bias.view(1, 1, self.n_focus, self.out_channels)
+            out = torch.cat([out[:, :1] + bias, out[:, 1:]], dim=1)
+        return out
+
+    def __repr__(self) -> str:
+        return (
+            f"{self.__class__.__name__}(lmax={self.lmax}, "
+            f"in_channels={self.in_channels}, out_channels={self.out_channels}, "
+            f"n_focus={self.n_focus})"
         )
